changes rename, abstract
rename ParamaR3
abstract okhttpHandler
changes .ignore
