+++ .gitignore
@@ -83,3 +83,10 @@ lint/generated/
 lint/outputs/
 lint/tmp/
 # lint/reports/
+
+cashes/
+.tmp/
+wrapper/
+native/
+android/
+daemon/
