+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:name="com.google.android.actions"
         android:resource="@xml/network_security_config" />
     <activity
-        android:name=".MainActivity"
+        android:name=".R3"
         android:exported="true">
         <intent-filter>
             <action android:name="android.intent.action.MAIN" />
